+++ db/pvault.sql
@@ -210,6 +210,9 @@ CREATE TABLE `ingredients` (
   `cat11B` varchar(255) COLLATE utf8_bin DEFAULT NULL,
   `cat12` varchar(255) COLLATE utf8_bin DEFAULT NULL,
   `manufacturer` varchar(255) COLLATE utf8_bin DEFAULT NULL,
+  `impact_top` varchar(10) COLLATE utf8_bin DEFAULT NULL,
+  `impact_heart` varchar(10) COLLATE utf8_bin DEFAULT NULL,
+  `impact_base` varchar(10) COLLATE utf8_bin DEFAULT NULL,  
   `created` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_bin;