+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - ADD: Upload ingredients to PV Online
 - ADD: Quantity of product to purchase in cart page
 - ADD: Add LOG/P in ingredients
+- ADD: Advanced note impact in ingredients page
 - PRV: Option to exclude notes when uploading ingredients to PV Online
 - CHG: Moving Olfactory Pyramid view to a tab in the formula page
 - FIX: Not showing allergens correctly in label